Update README.md
BruteWallet is an innovative tool designed to assist in the recovery of lost or forgotten cryptocurrency wallets. This program utilizes powerful brute-force algorithms to regain access to wallets by trying a vast array of possible key combinations. BruteWallet supports various popular cryptocurrency wallets and offers an easy-to-use interface, allowing users to conveniently configure search parameters, such as the type of crypto wallet and likely components of the forgotten or misplaced password. Ensuring user data protection is a priority, hence all recovery processes occur on local devices without the need to transmit private information online. BruteWallet is an indispensable tool for anyone looking to reclaim access to their cryptocurrency investments.